% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware.R
\name{scilis}
\alias{scilis}
\title{Scilis}
\usage{
scilis(secret)
}
\arguments{
\item{secret}{Secret for has creation.
Ideally use an environment variable so as to keep
this secret.}
}
\description{
Signs (and unsigns) cookies to ensure they are not
altered by clients.
}
\details{
This signs the values of cookies with
a base 64 encoded sha256 hash. Note that the
value of the cookie is still visible.
When cookies are being read the value is compared
to the expected hash to ensure it was not altered
by the client.
If the cookie appears aletered the returned value
is \code{character(0L)}.
}
\examples{
app <- ambiorix::Ambiorix$
 new()$
 use(scilis("secret"))

}
